requirements.txt
Flask==3.0.0
RPi.GPIO==0.7.1
Flask-CORS==4.0.0
python-dotenv==1.0.0
Flask-SocketIO==5.3.6
